новый файл:    apk_list 	изменено:      install_apk.sh

+++ install_apk.sh
@@ -1,2 +1,15 @@
-curl -L -o apk.tar.xz     http://github.com/definitly486/Lenovo_Tab_3_7_TB3-730X/releases/download/apk/apk.tar.xz   
-tar xf apk.tar.xz 
+
+if ! [  -f "apk.tar.xz" ]; then
+     curl -L -o apk.tar.xz     http://github.com/definitly486/Lenovo_Tab_3_7_TB3-730X/releases/download/apk/apk.tar.xz   
+fi
+
+if ! [  -f "apk" ]; then
+     tar xf apk.tar.xz 
+fi 
+
+
+  for i in $(cat apk_list ); do
+
+             pm install  $i 
+
+done
